nas-server/requirements.txt
Flask
PyJWT
python-dotenv